+++ src/Service/Parser/StringParser.php
@@ -5,28 +5,16 @@ namespace FD\LogViewer\Service\Parser;
 
 class StringParser
 {
-    public const WHITESPACE = [' ' => true, "\t" => true, "\n" => true, "\r" => true];
-
-    public function __construct(private QuotedStringParser $quotedStringParser)
+    public function __construct(private readonly QuotedStringParser $quotedStringParser, private readonly WordParser $wordParser)
     {
     }
 
     public function parse(StringReader $string): string
     {
-        $char = $string->peek();
-        if ($char === '"' || $char === "'") {
-            return $this->quotedStringParser->parse($string, $char, '\\');
+        if (in_array($string->get(), ['"', "'"], true)) {
+            return $this->quotedStringParser->parse($string, $string->get(), '\\');
         }
 
-        $result = '';
-        for (; $string->eol() === false; $string->next()) {
-            $char = $string->get();
-            if (isset(self::WHITESPACE[$char])) {
-                break;
-            }
-            $result .= $char;
-        }
-
-        return $result;
+        return $this->wordParser->parse($string);
     }
 }
