+++ tests/Unit/Service/Parser/StringParserTest.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace FD\LogViewer\Tests\Unit\Service\Parser;
+
+use FD\LogViewer\Service\Parser\StringParser;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\TestCase;
+
+#[CoversClass(StringParser::class)]
+class StringParserTest extends TestCase
+{
+    public function testParse(): void
+    {
+    }
+}
